backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
langchain
langchain-community
sentence-transformers
faiss-cpu
google-generativeai
numpy
pydantic
python-dotenv
